refactor(llm): extract shared retry helpers to llm::retry module
NoRetry, is_retryable, backoff_duration, MAX_RETRIES, and
BACKOFF_BASE_MS are identical for Anthropic and OpenAI providers.
Extract to a shared crate-private module to avoid duplication.
Moved retry tests from anthropic::tests to retry::tests.
