+++ src/llm/mod.rs
@@ -3,4 +3,5 @@
 pub mod anthropic;
 pub mod openai;
 pub mod provider;
+pub(crate) mod retry;
 pub mod types;
